refactor(BA-6600): extract Suggestion/SuggestionKind into selectors.types
Move the remediation value types (Suggestion, SuggestionKind) into a
dedicated selectors/types.py that depends only on common value types, so
exceptions.py can return a Suggestion without importing selection logic.
The selection result types (RequirementSelectionFailure,
BatchAgentSelectionResult) stay in selector.py since they depend on
AgentSelection/ResourceRequirements defined there.
